[#6] change package name inventory2logistics
패키지 이름을 inventory에서 logistics로 변경함.
